Refactor to use vulkan c++ pt.3. Working
